requirements.txt
anthropic
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
requests
python-dotenv
pyjwt
Pillow
